api relocate
